+++ gendox-frontend/src/views/apps/chat/components/chat-content/SendMsgForm.js
