+++ gendox-frontend/src/views/apps/chat/components/chat-content/UserProfileRight.js
