+++ gendox-frontend/src/views/apps/chat/components/chat-content/chatLog-components/ChatLogScrollWrapper.js
