+++ gendox-frontend/src/views/apps/chat/components/chat-content/chatLog-components/chatLog-message-group/chatLog-message-components/ChatLogInfo.js
